--- src/model/funcs/dateTime.ts
@@ -1,9 +1,0 @@
-import { getNextSunday, now } from "../../pkg/dayjs";
-
-const getStartAndEndUtcDateTime = () => {
-  const dateNowUtc = now();
-  const nextSundayUtc = getNextSunday(dateNowUtc).format();
-  return { dateNowUtc, nextSundayUtc };
-};
-
-export { getStartAndEndUtcDateTime };
