+++ src/model/user/user_model.ts
@@ -1,8 +1,8 @@
 import { Prisma } from "@prisma/client";
 import { prisma } from "../../db/db";
 import { PrismaClientWithTx } from "../../db/types";
-import { getUpdatedLevelAndExp } from "../funcs/exp";
 import { User } from "./types";
+import { getUpdatedLevelAndExp } from "../../funcs/exp";
 
 export class UserModel {
   public async getById(id: string): Promise<User | null> {
